refactor todo reducers to be functions
