+++ examples/todos/common/src/commonMain/kotlin/org/reduxkotlin/examples/todos/VisibilityFilterReducer.kt
@@ -1,0 +1,8 @@
+package org.reduxkotlin.examples.todos
+
+
+fun visibilityFilterReducer(state: VisibilityFilter, action: Any) =
+    when (action) {
+        is SetVisibilityFilter -> action.visibilityFilter
+        else -> state
+    }
